converted Settings page and its subcomponents from jsx to tsx
