+++ backlog.txt
@@ -14,6 +14,7 @@ add "assign composition to other composer" feature
 add "remove" button for composer image
 add MacOS support
 add more GUI languages
+add midi port in settings page
 
 Bugs:
 =====
