+++ www/client/package.json
@@ -57,6 +57,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react-color": "^3.0.13",
+    "@types/react-slider": "^1.3.6",
     "@vitejs/plugin-react": "^5.0.1",
     "@vitejs/plugin-react-swc": "^4.0.1",
     "cross-env": "^7.0.3",
